+++ wpdk.php
@@ -262,7 +262,7 @@ if ( !class_exists( 'WPDK' ) ) {
         // USER INTERFACE
         //------------------------------------------------------------------
 
-        $sPathPrefix . 'classes/ui/wpdk-view-controller.php'               => array(
+        $sPathPrefix . 'classes/ui/wpdk-viewcontroller.php'               => array(
           'WPDKView',
           'WPDKViewController',
           'WPDKHeaderView',
@@ -313,6 +313,8 @@ if ( !class_exists( 'WPDK' ) ) {
           'WPDKUI'
         ),
 
+        $sPathPrefix . 'classes/ui/wpdk-glyphicons.php'                    => 'WPDKGlyphIcons',
+
         $sPathPrefix . 'classes/ui/wpdk-twitter-bootstrap.php'             => array(
           'WPDKTwitterBootstrap',
           'WPDKTwitterBootstrapModal',
@@ -346,7 +348,7 @@ if ( !class_exists( 'WPDK' ) ) {
           'WPDKSubMenu',
           'WPDKSubMenuDivider',
         ),
-        $sPathPrefix . 'classes/ui/wpdk-listtable-vc.php'                  => 'WPDKListTableViewController',
+        $sPathPrefix . 'classes/ui/wpdk-listtable-viewcontroller.php'                  => 'WPDKListTableViewController',
         $sPathPrefix . 'classes/ui/wpdk-pointer.php'                       => 'WPDKPointer',
 
         //------------------------------------------------------------------